requirements.txt
config
psycopg2-binary
paho-mqtt